fix: remove kafka dependency
Removed kafka-python from dependencies with the command 'poetry remove kafka', which modified the poetry.lock and pyproject.toml files. In the .env file, I removed the ports for some kafka server. Two tests for the annotation/annotation/main.py module are skipped as they use a function with 'kafka' in their title.